hotfix: 티켓 검색 응답 오류 수정 (#121)
* feat: 티켓 검색 응답 오류 수정

* feat: remove unuseless log
