+++ src/main/kotlin/uket/api/admin/AdminTicketController.kt
@@ -37,7 +37,6 @@ import uket.domain.uketevent.service.UketEventService
 import uket.facade.EnterUketEventFacade
 import uket.facade.TicketEntryGroupUserFacade
 import uket.facade.UpdateTicketStatusFacade
-import uket.facade.assembler.TicketSearchAssembler
 import uket.facade.search.TicketSearcher
 
 @Tag(name = "어드민 티켓 관련 API", description = "어드민 티켓 관련 API 입니다.")
@@ -52,7 +51,6 @@ class AdminTicketController(
     private val adminService: AdminService,
     private val uketEventService: UketEventService,
     private val ticketEntryGroupUserFacade: TicketEntryGroupUserFacade,
-    private val ticketSearchAssembler: TicketSearchAssembler,
 ) {
     @SecurityRequirement(name = "JWT")
     @Operation(summary = "입장 확인 API", description = "QR code를 통한 Token값으로 입장 확인을 할 수 있습니다.")
@@ -128,7 +126,7 @@ class AdminTicketController(
                             title = "잘못된 검색 타입"
                         )
                     }.search(adminInfo.organizationId, uketEventId, searchRequest, pageRequest);
-                ticketSearchAssembler.toDtoPage(tickets)
+                ticketEntryGroupUserFacade.toDtoPage(tickets)
             }
 
         val response = CustomPageResponse(TicketSearchResponse.from(tickets))
